# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /characters:
    get:
      operationId: listCharacters
      tags: [characters]
      summary: List characters
      responses:
        "200":
          description: Character library
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Character"
    post:
      operationId: createCharacter
      tags: [characters]
      summary: Create a character
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CharacterInput"
      responses:
        "201":
          description: Created character
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Character"
  /characters/{id}:
    patch:
      operationId: updateCharacter
      tags: [characters]
      summary: Update a character
      parameters:
        - $ref: "#/components/parameters/Id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CharacterInput"
      responses:
        "200":
          description: Updated character
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Character"
    delete:
      operationId: deleteCharacter
      tags: [characters]
      summary: Delete a character
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "204":
          description: Deleted
  /characters/{id}/generate-image:
    post:
      operationId: generateCharacterImage
      tags: [characters]
      summary: Generate and attach a character reference image
      parameters:
        - $ref: "#/components/parameters/Id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StudioImageGenerationInput"
      responses:
        "201":
          description: Generated character reference image
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudioImageGenerationResult"
  /characters/{id}/voice-sample:
    post:
      operationId: uploadCharacterVoiceSample
      tags: [characters]
      summary: Upload a consented character voice reference
      parameters:
        - $ref: "#/components/parameters/Id"
      requestBody:
        required: true
        content:
          audio/wav:
            schema: { type: string, format: binary }
          audio/mpeg:
            schema: { type: string, format: binary }
          audio/mp4:
            schema: { type: string, format: binary }
          audio/webm:
            schema: { type: string, format: binary }
      responses:
        "201":
          description: Normalized voice reference
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CharacterVoiceSampleResult"
    delete:
      operationId: deleteCharacterVoiceSample
      tags: [characters]
      summary: Remove a character voice reference
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "204":
          description: Voice reference removed
...
components:
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema:
        type: string
  schemas:
...
    Character:
      type: object
      required: [id, name, description, promptDescription, thumbnail, tags, assetCount, voiceProfile, hasVoiceSample, voiceSampleUrl, voiceConsentAt, createdAt, updatedAt]
      properties:
        id: { type: string }
        name: { type: string }
        description: { type: string }
        promptDescription: { type: string }
        thumbnail: { type: ["string", "null"] }
        tags: { type: array, items: { type: string } }
        assetCount: { type: number }
        voiceProfile: { type: ["string", "null"] }
        hasVoiceSample: { type: boolean }
        voiceSampleUrl: { type: ["string", "null"] }
        voiceConsentAt: { type: ["string", "null"] }
        createdAt: { type: string }
        updatedAt: { type: string }
    CharacterInput:
      type: object
      required: [name, description, promptDescription]
      properties:
        name: { type: string, minLength: 1 }
        description: { type: string }
        promptDescription: { type: string }
        thumbnail: { type: ["string", "null"] }
        tags: { type: array, items: { type: string } }
        voiceProfile: { type: ["string", "null"] }
    CharacterVoiceSampleResult:
      type: object
      required: [ok, voiceSampleUrl, voiceConsentAt]
      properties:
        ok: { type: boolean }
        voiceSampleUrl: { type: string }
        voiceConsentAt: { type: string }
...
    StudioImageGenerationInput:
      type: object
      properties:
        prompt: { type: string, minLength: 1, maxLength: 4000 }
        seed: { type: number, minimum: 0, maximum: 2147483647 }
    StudioImageGenerationResult:
      type: object
      required: [ok, assetId, mediaUrl, serverName, seed]
      properties:
        ok: { type: boolean }
        assetId: { type: string }
        mediaUrl: { type: string }
        serverName: { type: string }
        seed: { type: number }
...